requirements.txt
asgiref==3.3.4
blis==0.7.4
catalogue==2.0.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
cymem==2.0.5
dj-config-url==0.1.1
Django==3.2.2
django-extensions==3.1.3
djangorestframework==3.12.4
drf-yasg==1.20.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
environ==1.0
idna==2.10
importlib-metadata==4.0.1
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
joblib==1.0.1
langdetect==1.0.9
MarkupSafe==1.1.1
murmurhash==1.0.5
nltk==3.6.2
numpy==1.20.3
packaging==20.9
pathy==0.5.2
preshed==3.0.5
pydantic==1.7.4
pydotplus==2.0.2
pymystem3==0.2.0
pyparsing==2.4.7
pytz==2021.1
regex==2021.4.4
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
six==1.16.0
smart-open==3.0.0
spacy==3.0.6
spacy-legacy==3.0.5
sqlparse==0.4.1
srsly==2.4.1
thinc==8.0.3
tqdm==4.60.0
typer==0.3.2
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.4
wasabi==0.8.2
zipp==3.4.1
